Adding a confirmation before deleting any post or comment , creation of the css file for the dashboard

+++ public/js/deletePost.js
@@ -1,16 +1,34 @@
 $(function () {
-    $(".delete-post").on("click", (e) => {
-        e.preventDefault();
-        $.ajax({
-            type: 'DELETE',
-            url: e.target.href,
-            success: function(response){
-                let test = JSON.parse(response);
 
-                if (test === 'test') {
-                    $(e.target).parent().remove()
+    $(".delete-post, .delete-com").on("click", function (e)  {
+        e.preventDefault();
+        console.log(e);
+        let targetEvent = $(e.target).parent().parent();
+        let deleteClass = "";
+
+        if (e.target.href.includes("post")){
+            deleteClass = "cet article";
+        } else {
+            deleteClass = "ce commentaire";
+        }
+
+        let choice = confirm('Voulez vous vraiment effacer ' + deleteClass);
+
+        if (choice){
+            $.ajax({
+                type: 'DELETE',
+                url: e.target.href,
+                success: function (response) {
+                    let test = JSON.parse(response);
+                    if (test === 'success') {
+                        targetEvent.fadeTo(600, 0.00, function(){
+                            targetEvent.slideUp(400, function() {
+                                targetEvent.remove();
+                            });
+                        });
+                    }
                 }
-            }
-        });
+            });
+        }
     });
 });
